+++ commander/src/HorusConfig/difficulty_subpage.h
@@ -1,12 +1,14 @@
 #ifndef DIFFICULTY_SUBPAGE_H
 #define DIFFICULTY_SUBPAGE_H
 
+#include <QObject>
+
 class DifficultySubpage
 {
 public:
     DifficultySubpage();
-    virtual int getDifficultyCode() = 0;
-    virtual void setDifficultyCode(int value) = 0;
+    virtual qint64 getDifficultyCode() = 0;
+    virtual void setDifficultyCode(qint64 value) = 0;
 };
 
 #endif // DIFFICULTY_SUBPAGE_H
